Add files via upload
The Loan Approval Data Science project aims to build a predictive model to streamline and automate the loan approval process for financial institutions. The objective is to accurately predict whether an applicant should be approved for a loan based on their demographic and financial features. By leveraging machine learning techniques, this project seeks to reduce the time and effort required for manual assessments while improving the accuracy of loan approval decisions.